+++ CHANGELOG.rst
@@ -1,9 +1,17 @@
 Unreleased changes
 ------------------
+
+
+4.8.2 / 2016-10-13
+------------------
+* Fixes geometry application in HitSeries
+
+4.8.1 / 2016-10-12
+------------------
 * Forcing matplotlib 2.0.0b4 as dependency. Don't blame us!
 * New unified style for all plots, using `import km3pipe.style`
 * aanet / jgandalf: write zeroed row if no track in event
-
+* fix string handling in H5 attributes
 
 4.8.0 / 2016-10-11
 ------------------